backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
readability-lxml==0.8.1
weasyprint==62.3
requests==2.32.3
lxml==5.3.0
lxml_html_clean==0.2.2
